# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Faux Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api/v1"
        }
    ],
    "paths": {
...
        "/students/{student}": {
            "get": {
                "tags": [
                    "Common data"
                ],
                "summary": "Student",
                "description": "Get student's data",
                "operationId": "studentItself",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "1": {
                                "summary": "Data of the student with ID=1",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Data array",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "Student's name",
                                                        "type": "string",
                                                        "example": "Domenica Kris"
                                                    },
                                                    "group": {
                                                        "description": "Student's group",
                                                        "type": "string",
                                                        "example": "AA-00"
                                                    },
                                                    "courses": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "": {
                                                                    "description": "Course name",
                                                                    "type": "string",
                                                                    "example": "Math"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "example": "Wrong request!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "get_student"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "Fail due to some errors"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Student delete",
                "description": "Delete student data",
                "operationId": "studentDelete",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "5": {
                                "summary": "Student ID",
                                "value": "5"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Data array",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "message": {
                                                        "description": "Result of the edit of the student data",
                                                        "type": "string",
                                                        "example": "Success"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Wrong request!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "delete_student'"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "Fail due to some errors"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "delete_student'"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "Fail due to some errors"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Student edit",
                "description": "Edit student data",
                "operationId": "studentEdit",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "Student ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "5": {
                                "summary": "Student ID",
                                "value": "5"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "group_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Domenica"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Denesik"
                                            },
                                            "course_ids": {
                                                "type": "array",
                                                "items": {
                                                    "description": "Course ID",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Data array",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "message": {
                                                        "description": "Result of the edit of the student data",
                                                        "type": "string",
                                                        "example": "Success"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Wrong request!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "edit_student'"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "Fail due to some errors"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "edit_student'"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "Fail due to some errors"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}